Bugfix in waitForFile where fn was full file path
This was identified when initialising jobs on a slurm cluster when called from readLog  with fs.latency = 65. fn is a full file path, so isn't found in list.files(path, all.files = TRUE) without the full.names argument. Rather than adding the full.names argument to list.files, this solution handles fn as either a full file path or a file name, and provides the full file path in the timeout error message.